+++ tests/testthat/test-listings.R
@@ -1,7 +1,7 @@
 testthat::test_that("Column labels are the same", {
 
   ## listings var labels don't get mucked up by topleft machinery #262
-  lsting <- as_listing(anl, key_cols = c("USUBJID")) %>%
+  lsting <- as_listing(anl, key_cols = c("USUBJID"), cols = NULL) %>%
     add_listing_col("ARM")
 
   testthat::expect_identical(var_labels(anl), var_labels(lsting))
@@ -25,7 +25,7 @@ testthat::test_that("listings work well with different formats and attributes", 
   var_labels(anl_tmp) <- var_labels(ex_adsl)[c("USUBJID", "ARM", "BMRKR1")]
   anl_tmp$BMRKR1[1:3] <- NA
 
-  lsting <- as_listing(anl_tmp, key_cols = c("ARM", "USUBJID")) %>%
+  lsting <- as_listing(anl_tmp, key_cols = c("ARM", "USUBJID"), cols = NULL) %>%
     add_listing_col("ARM") %>%
     add_listing_col("USUBJID") %>%
     add_listing_col("BMRKR1", format = "xx.xx")
@@ -103,7 +103,7 @@ testthat::test_that("Content of listings supports newlines", {
 
   mydf <- data.frame(col1 = c(1, 1, 2), col2 = c("hi\nthere", "what", "who"))
   var_labels(mydf) <- c("column\n1", "column 2")
-  mylst <- as_listing(mydf, names(mydf), key_cols = "col1")
+  mylst <- as_listing(mydf, cols = names(mydf), key_cols = "col1")
 
   mpf <- matrix_form(mylst)
 
@@ -131,4 +131,43 @@ testthat::test_that("regression test for keycols being lost due to `head()`", {
     dim(head(rlst, 5)),
     c(5L, ncol(mtcars))
   )
+})
+
+testthat::test_that("default values work with as_listing", {
+  iiris <- iris[1:15, ]
+
+  # All the columns, first as key_cols
+  lst <- as_listing(iiris[, 1:2])
+  testthat::expect_equal(ncol(lst), 2L)
+  testthat::expect_equal(nrow(lst), 15L)
+  testthat::expect_true(any(matrix_form(lst)$strings[, 1] == "")) # First is key_cols
+  testthat::expect_equal(matrix_form(lst)$strings[, 1][1], "Sepal.Length")
+
+  # Selecting other key_cols keeps all the listing
+  lst <- as_listing(iiris, key_cols = "Species")
+  testthat::expect_equal(ncol(lst), 5L)
+  testthat::expect_equal(nrow(lst), 15L)
+  testthat::expect_true(all(matrix_form(lst)$strings[c(-1, -2), 1] == "")) # First is key_cols
+  testthat::expect_equal(matrix_form(lst)$strings[, 1][1], "Species")
+
+  # Selecting disjointed key_cols and cols renders them both
+  lst <- as_listing(iiris, cols = "Sepal.Width", key_cols = "Species")
+  # testthat::expect_equal(ncol(lst), 2L) # Error? to fix
+  testthat::expect_equal(nrow(lst), 15L)
+  testthat::expect_true(all(matrix_form(lst)$strings[c(-1, -2), 1] == "")) # First is key_cols
+  testthat::expect_equal(matrix_form(lst)$strings[, 1][1], "Species")
+
+  # Using NULL for cols shows no other columns beside key_cols
+  lst <- as_listing(iiris, cols = NULL, key_cols = "Species")
+  # testthat::expect_equal(ncol(lst), 1L) # Error? to fix
+  testthat::expect_equal(nrow(lst), 15L)
+  testthat::expect_true(all(matrix_form(lst)$strings[c(-1, -2), 1] == "")) # First is key_cols
+  testthat::expect_equal(matrix_form(lst)$strings[, 1][1], "Species")
+
+  # Using both NULLs for cols and key_cols produces warning and prints first column
+  lst <- testthat::expect_warning(as_listing(iiris, cols = NULL))
+  # testthat::expect_equal(ncol(lst), 1L) # Error? to fix
+  testthat::expect_equal(nrow(lst), 15L)
+  testthat::expect_true(any(matrix_form(lst)$strings[, 1] == "")) # First is key_cols
+  testthat::expect_equal(matrix_form(lst)$strings[, 1][1], "Sepal.Length")
 })